requirements.txt
# AI E-commerce Product Scout - Dependencies
# ============================================

# Agent Framework
google-adk>=0.3.0
google-genai>=1.0.0

# Backend API
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0

# Frontend
streamlit>=1.40.0

# Database
psycopg2-binary>=2.9.9

# HTTP Client (for Streamlit -> FastAPI)
requests>=2.32.0

# Environment
python-dotenv>=1.0.0
